+++ src/main/java/com/swyp/artego/domain/commission/repository/CommissionRepository.java
@@ -1,6 +1,6 @@
-package com.swyp.artego.domain.code.commission.repository;
+package com.swyp.artego.domain.commission.repository;
 
-import com.swyp.artego.domain.code.commission.entity.Commission;
+import com.swyp.artego.domain.commission.entity.Commission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommissionRepository extends JpaRepository<Commission, Long> {
