+++ src/main/java/com/swyp/artego/domain/follow/entity/Follow.java
@@ -1,6 +1,6 @@
-package com.swyp.artego.domain.user.entity;
+package com.swyp.artego.domain.follow.entity;
 
-import com.swyp.artego.domain.post.entity.Item;
+import com.swyp.artego.domain.user.entity.User;
 import com.swyp.artego.global.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -8,30 +8,27 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "scrap")
-public class Scrap extends BaseTimeEntity {
+@Table(name = "follow")
+public class Follow extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "scrap_id", nullable = false)
     private Long id;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="user_id")
     private User user;
 
-    @ManyToOne
-    @JoinColumn(name="item_id")
-    private Item item;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="user_artist_id")
+    private User userArtist;
 
     @Builder
-    public Scrap(User user, Item item) {
+    public Follow(User user, User userArtist) {
         this.user = user;
-        this.item = item;
+        this.userArtist = userArtist;
     }
 }
