+++ src/main/java/com/swyp/artego/domain/follow/repository/FollowRepository.java
@@ -1,6 +1,6 @@
-package com.swyp.artego.domain.user.repository;
+package com.swyp.artego.domain.follow.repository;
 
-import com.swyp.artego.domain.user.entity.Follow;
+import com.swyp.artego.domain.follow.entity.Follow;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FollowRepository extends JpaRepository<Follow, Long> {
