+++ src/main/java/com/swyp/artego/domain/itemEmoji/entity/ItemEmoji.java
@@ -1,6 +1,7 @@
-package com.swyp.artego.domain.user.entity;
+package com.swyp.artego.domain.itemEmoji.entity;
 
-import com.swyp.artego.domain.post.entity.Item;
+import com.swyp.artego.domain.item.entity.Item;
+import com.swyp.artego.domain.user.entity.User;
 import com.swyp.artego.global.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
