+++ src/main/java/com/swyp/artego/domain/scrap/repository/ScrapRepository.java
@@ -1,6 +1,6 @@
-package com.swyp.artego.domain.user.repository;
+package com.swyp.artego.domain.scrap.repository;
 
-import com.swyp.artego.domain.user.entity.Scrap;
+import com.swyp.artego.domain.scrap.entity.Scrap;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ScrapRepository extends JpaRepository<Scrap, Long> {
