--- src/main/java/com/swyp/artego/domain/user/entity/Follow.java
@@ -1,33 +1,0 @@
-package com.swyp.artego.domain.user.entity;
-
-import com.swyp.artego.global.common.entity.BaseTimeEntity;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "follow")
-public class Follow extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="user_id")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="user_artist_id")
-    private User userArtist;
-
-    @Builder
-    public Follow(User user, User userArtist) {
-        this.user = user;
-        this.userArtist = userArtist;
-    }
-}
